smart_med_api/requirements.txt
fastapi
httpx
firebase-admin
openai
pydantic
python-dotenv
python-multipart
uvicorn[standard]
